+++ core/src/main/java/org/jdbi/v3/core/result/ResultBearing.java
@@ -36,6 +36,7 @@ import org.jdbi.v3.core.mapper.RowViewMapper;
 import org.jdbi.v3.core.mapper.SingleColumnMapper;
 import org.jdbi.v3.core.mapper.reflect.BeanMapper;
 import org.jdbi.v3.core.qualifier.QualifiedType;
+import org.jdbi.v3.core.result.internal.ResultSetResultIterable;
 import org.jdbi.v3.core.result.internal.RowViewImpl;
 import org.jdbi.v3.core.statement.StatementContext;
 
@@ -129,7 +130,7 @@ public interface ResultBearing {
         return scanResultSet((resultSetSupplier, ctx) -> {
             RowMapper<T> rowMapper = ctx.findMapperFor(type)
                     .orElseThrow(() -> new NoSuchMapperException("No mapper registered for type " + type));
-            return ResultIterable.of(resultSetSupplier, rowMapper, ctx);
+            return new ResultSetResultIterable<>(rowMapper, ctx, resultSetSupplier, type.getType());
         });
     }
 
@@ -164,7 +165,7 @@ public interface ResultBearing {
      */
     default <T> ResultIterable<Map<String, T>> mapToMap(Class<T> valueType) {
         return scanResultSet((resultSetSupplier, ctx) ->
-                ResultIterable.of(resultSetSupplier, GenericMapMapperFactory.getMapperForValueType(valueType, ctx.getConfig()), ctx));
+                new ResultSetResultIterable<>(GenericMapMapperFactory.getMapperForValueType(valueType, ctx.getConfig()), ctx, resultSetSupplier, Object.class));
     }
 
     /**
@@ -177,7 +178,7 @@ public interface ResultBearing {
      */
     default <T> ResultIterable<Map<String, T>> mapToMap(GenericType<T> valueType) {
         return scanResultSet((resultSetSupplier, ctx) ->
-                ResultIterable.of(resultSetSupplier, GenericMapMapperFactory.getMapperForValueType(valueType, ctx.getConfig()), ctx));
+                new ResultSetResultIterable<>(GenericMapMapperFactory.getMapperForValueType(valueType, ctx.getConfig()), ctx, resultSetSupplier, Object.class));
     }
 
     /**
@@ -199,7 +200,7 @@ public interface ResultBearing {
      * @return a {@link ResultIterable} of type {@code <T>}.
      */
     default <T> ResultIterable<T> map(RowMapper<T> mapper) {
-        return scanResultSet((resultSetSupplier, ctx) -> ResultIterable.of(resultSetSupplier, mapper, ctx));
+        return scanResultSet((resultSetSupplier, ctx) -> new ResultSetResultIterable<>(mapper, ctx, resultSetSupplier, Object.class));
     }
 
     /**
@@ -399,7 +400,7 @@ public interface ResultBearing {
             RowMapper<?> rowMapper = ctx.findMapperFor(elementType)
                     .orElseThrow(() -> new NoSuchMapperException("No mapper registered for element type " + elementType));
 
-            return ResultIterable.of(resultSetSupplier, rowMapper, ctx).collect(collector);
+            return new ResultSetResultIterable<>(rowMapper, ctx, resultSetSupplier, elementType).collect(collector);
         });
     }
 }